+++ docs/change_log.rst
@@ -3,3 +3,4 @@
 - Refactored Uncertainty.make_samples_from_distributions(..) to Uncertainty.sample_distributions(..)
 - Added Uncertainty.sample_uniforms
 - Added test_simple_model
+- Added MixedInhibition2, which takes kic and kiu